+++ CHANGELOG.md
@@ -13,8 +13,11 @@
 
 ### Deprecated
 
+Deprecated functions may be removed in a future build
+
 - `isEven` in favor of just `even`
 - `isOdd` in favor of just `odd`
+- `isPrime` in favor of just `prime`
 
 ### New